# feat: optimise local diff