--- a/.env.test
+++ b/.env.test
@@ -35,7 +35,7 @@ export E2E_TEST_SECRET="test"
 export E2E_TEST_GH_TOKEN="test"
 
 # Database
-export DB_URI="postgres://isomer:password@localhost:54321/isomercms_test"
+export DB_URI="postgres://isomer:password@127.0.0.1:54321/isomercms_test"
 export DB_MIN_POOL="1"
 export DB_MAX_POOL="10"
 export DB_ENABLE_LOGGING="true"